dev-requirements.txt
langdetect==1.0.9
logger_setup==0.17
mysql_connector_repackaged==0.3.1
numpy==2.2.6
python-dotenv==1.1.0
python_Levenshtein==0.27.1
PyYAML==6.0.2
Requests==2.32.3
scikit_learn==1.6.1
watchdog==6.0.0
pre-commit
black
docformatter
flake8
flake8-docstrings
flake8-bugbear
flake8-comprehensions
coverage
distlib
isort
packaging
platformdirs
pylint
pytest
tomlkit
virtualenv
vulture
pyupgrade
